requirements.txt
streamlit
langgraph
langchain-google-genai
langchain-community
chromadb
python-dotenv
requests
beautifulsoup4
trafilatura
pandas
numpy
tiktoken
lxml_html_clean
duckduckgo-search


